pyproject.toml
[tool.poetry]
name = "nest"
version = "0.1.0"
description = "Noi Estraiamo Statistiche Tweet"
authors = [
    "Lorenzo Balugani <256867@studenti.unimore.it>",
    "Stefano Pigozzi <256895@studenti.unimore.it>",
    "Stefano Goldoni <280712@studenti.unimore.it>",
    "Giorgio Minoccari <261807@studenti.unimore.it>",
    "Flavia Cocca <253218@studenti.unimore.it>",
    "Chiara Calzolari <258727@studenti.unimore.it>",
    "Giovanni Anniballi <253150@studenti.unimore.it>",
]
packages = [
    { include = "nest_backend" },
    { include = "nest_crawler" },
]

[tool.poetry.dependencies]
python = "^3.8.5"
psycopg2-binary = "^2.8.6"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
bcrypt = "^3.2.0"
Flask-JWT-Extended = "^4.1.0"
Flask-Cors = "^3.0.10"
requests = "^2.25.1"
apispec = "^4.4.0"
apispec-webframeworks = "^0.5.2"
marshmallow = "^3.11.1"
openapi-spec-validator = "^0.3.0"
flask-swagger-ui = "^3.36.0"
tweepy = "^3.10.0"
nltk = "^3.6.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
Sphinx = "^3.5.4"
sphinx-rtd-theme = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
